requirements.txt
# Web crawling
requests
beautifulsoup4
selenium

# Data handling
tqdm
numpy
pandas

# Embedding + NLP
sentence-transformers
transformers
torch

# Vector Store
chromadb
langchain
langchain-community

# LLMs
openai
google-generativeai

# Utils
joblib
python-dotenv

# Computer Vision
ultralytics
Pillow

# Web App
streamlit
requests

# Database
python-jose[cryptography]
passlib[bcrypt]


pydantic[email]

streamlit-autorefresh

#map
folium
streamlit-folium